Refactoring code
* Extract methods to ntjoin_utils.py helper script
